moved around files and split css file
