+++ Front-End/src/Components/Other-Pages/Bug.jsx
