+++ Front-End/src/css/Gamepage.css
@@ -1,0 +1,102 @@
+.video {
+    z-index: 1;
+  }
+  
+  .youtube-wrapper {
+    display: flex;
+    justify-content: center;
+  }
+  
+  @media screen and (max-width: 768px) {
+    .youtube-wrapper {
+      position: relative;
+      overflow: hidden;
+      justify-content: center;
+      padding-top: 56.25%;
+      margin-left: 0;
+    }
+  
+    .video {
+      position: absolute;
+      z-index: 1;
+      top: 0;
+      left: 0;
+      width: 100%;
+      height: 100%;
+    }
+  }
+  
+  .game-container {
+    margin-left: 5rem;
+  }
+  
+  .ranks {
+    display: flex;
+    flex-wrap: wrap;
+    justify-content: center;
+    align-items: center;
+    margin-top: 20px;
+    margin-bottom: 30px;
+  }
+  
+  .rank {
+    width: 100px;
+    height: 100px;
+    border-radius: 10rem;
+  }
+  
+  @media screen and (max-width: 768px) {
+    .rank {
+      width: 40px;
+      height: 40px;
+      border-radius: 10rem;
+    }
+  }
+  
+  .rank:hover {
+    transform: scale(1.1);
+    transition: all 0.3s ease-in-out;
+  }
+  
+  .immortal,
+  .iron,
+  .gold,
+  .silver,
+  .plat,
+  .diamond,
+  .bronze,
+  .radiant,
+  .asc {
+    margin: 0;
+    cursor: pointer;
+    border-radius: 3rem;
+  }
+  
+  .silver,
+  .plat {
+    margin-right: 10px;
+  }
+  
+  .diamond,
+  .gold,
+  .asc,
+  .immortal {
+    margin-right: 15px;
+  }
+  
+  .radiant {
+    margin-left: 15px;
+  }
+  
+  .submit {
+    cursor: pointer;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    border: none;
+    margin: 0 auto;
+    width: fit-content;
+    background-color: #1f1d1d;
+    font-size: 2rem;
+    color: white;
+  }
