+++ Front-End/src/css/Homepage.css
@@ -1,0 +1,55 @@
+
+.main-page-body {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    margin-top: 5rem;
+  }
+  
+  .page-heading-1 {
+    text-align: center;
+    font-weight: 500;
+    text-transform: uppercase;
+    letter-spacing: .2em;
+    color: rgb(255 255 255);
+  }
+  
+  @media (min-width: 1024px) {
+    .page-heading-1 {
+        font-size: 1.875rem;
+        line-height: 2.25rem;
+    }
+  }
+  
+  .main-images {
+    height: auto;
+    width: 12.5%;
+    margin-right: 1.5rem;
+  }
+  
+  .main-images:hover {
+    transform: scale(1.1);
+    transition: all 0.3s ease-in-out;
+  }
+  
+  .gamecard {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    justify-content: center;
+    padding: 20px;
+    text-align: center;
+  }
+  
+  @media screen and (max-width: 768px) {
+    .main-images {
+      height: auto;
+      width: 25%;
+      display: flex;
+      justify-content: center;
+      align-items: center;
+      margin: 0 auto;
+      margin-bottom: 1rem;
+    }
+  }
