+++ Front-End/src/css/Navbar.css
@@ -169,196 +169,7 @@ header {
   }
 }
 
-.main-page-body {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  margin-top: 5rem;
-}
-
-.page-heading-1 {
-  text-align: center;
-  font-weight: 500;
-  text-transform: uppercase;
-  letter-spacing: .2em;
-  color: rgb(255 255 255);
-}
-
-@media (min-width: 1024px) {
-  .page-heading-1 {
-      font-size: 1.875rem;
-      line-height: 2.25rem;
-  }
-}
-/* 
-h1 {
-  color: white;
-  text-align: center;
-  font-size: 30px;
-}
-
-.main-images {
-  height: auto;
-  width: 12.5%;
-  margin-right: 1.5rem;
-}
-
-.main-images:hover {
-  transform: scale(1.1);
-  transition: all 0.3s ease-in-out;
-}
-
-.gamecard {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: center;
-  padding: 20px;
-  text-align: center;
-}
-
-@media screen and (max-width: 768px) {
-  .main-images {
-    height: auto;
-    width: 25%;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin: 0 auto;
-    margin-bottom: 1rem;
-  }
-}
-
-@media screen and (min-width: 769px) {
-  .sidebar {
-    position: absolute;
-    top: 0;
-    left: 0;
-    height: 100%;
-    width: 260px;
-    background-color: rgb(31, 29, 29);
-    overflow-x: hidden;
-    padding-top: 20px;
-  }
-}
-
-.video {
-  z-index: 1;
-}
-
-.youtube-wrapper {
-  display: flex;
-  justify-content: center;
-}
-
-@media screen and (max-width: 768px) {
-  .youtube-wrapper {
-    position: relative;
-    overflow: hidden;
-    justify-content: center;
-    padding-top: 56.25%;
-    margin-left: 0;
-  }
-
-  .video {
-    position: absolute;
-    z-index: 1;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-  }
-}
-
-.game-container {
-  margin-left: 5rem;
-}
-
-.ranks {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  align-items: center;
-  margin-top: 20px;
-}
-
-.rank {
-  width: 100px;
-  height: 100px;
-  border-radius: 10rem;
-}
-
-@media screen and (max-width: 768px) {
-  .rank {
-    width: 40px;
-    height: 40px;
-    border-radius: 10rem;
-  }
-}
-
-.rank:hover {
-  transform: scale(1.1);
-  transition: all 0.3s ease-in-out;
-}
-
-.immortal,
-.iron,
-.gold,
-.silver,
-.plat,
-.diamond,
-.bronze,
-.radiant,
-.asc {
-  margin: 0;
-  cursor: pointer;
-  border-radius: 3rem;
-}
-
-.silver,
-.plat {
-  margin-right: 10px;
-}
-
-.diamond,
-.gold,
-.asc,
-.immortal {
-  margin-right: 15px;
-}
-
-.radiant {
-  margin-left: 15px;
-}
-
-.submit {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  height: 100%;
-  background-color: #292929;
-  color: #fff;
-  font-size: 1.5rem;
-  border: none;
-  cursor: pointer;
-  margin-top: 2rem;
-}
-
-.close {
-  color: #aaa;
-  float: right;
-  font-size: 28px;
-  font-weight: bold;
-}
-
-.close:hover,
-.close:focus {
-  color: black;
-  text-decoration: none;
-  cursor: pointer;
-}
+/* Modal CSS */
 
 .modal-content {
   margin: 9% auto;
@@ -408,90 +219,7 @@ h1 {
   color: #01ccff;
 }
 
-.form-container {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  align-items: center;
-  margin-top: clamp(10px, 4vw, 30px);
-  width: clamp(220px, 95%, 1000px);
-}
-
-.form {
-  width: 400px;
-  padding: 20px;
-}
-
-.select-game {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-bottom: 1rem;
-  flex-direction: column;
-}
-
-.select {
-  background-color: white;
-  border-radius: 10px;
-  display: flex;
-  flex-direction: row;
-  font-size: clamp(0.9rem, 3vw, 1.3rem);
-  font-weight: 400;
-  height: clamp(30px, 4vw, 45px);
-  justify-content: flex-start;
-  max-width: 600px;
-}
-
-.form-input,
-option,
-select,
-textarea {
-  background-color: white;
-  border: 1px solid black;
-  border-radius: 5px;
-  box-sizing: border-box;
-  color: var(--input-text-color);
-  font-size: clamp(0.8rem, 3vw, 1.1rem);
-  font-weight: 400;
-  height: clamp(30px, 4vw, 45px);
-  margin: initial;
-  outline: none;
-  padding: clamp(1px, 1vw, 6px) clamp(3px, 2vw, 15px);
-  width: 100%;
-  margin-top: 1rem;
-}
-
-.form-input,
-option {
-  height: clamp(30px, 4vw, 45px);
-}
-
-.form-label,
-.text,
-.login-text,
-.error-message,
-.howto {
-  color: white;
-}
-
-.form-label {
-  margin-left: 5px;
-  margin-top: 1rem;
-}
-
-.text {
-  margin-bottom: 1rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-}
-
-.submission-list,
-.form-checkbox,
-.other-links {
-  margin-top: 1rem;
-}
+/* Leaderboard CSS  */
 
 .leaderboard-container {
   display: flex;
@@ -525,21 +253,6 @@ td,
   color: white;
 }
 
-.container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-direction: column;
-  height: 75vh;
-}
-
-.signup-atag {
-  color: #01ccff;
-}
-
-.error-message {
-  color: red;
-}
 
 .submit-btn {
   background-color: #01ccff;
@@ -661,6 +374,7 @@ td,
   justify-content: center;
   align-items: center;
   margin-bottom: 10px;
+  margin-top: 1rem;
 }
 
 .report-align {
@@ -682,4 +396,4 @@ td,
   border-radius: 5px;
   padding: clamp(1px, 1vw, 6px) clamp(3px, 2vw, 15px);
   width: fit-content;
-} */
+}
